chore: migrate i18next-parser to i18next-cli
