--- i18next-parser.config.js
@@ -1,32 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires, no-undef
-import { CustomJSONLexer } from './i18n-scripts/lexers.js';
-
-// eslint-disable-next-line no-undef
-export default {
-  sort: true,
-  createOldCatalogs: false,
-  keySeparator: false,
-  locales: ['en'],
-  namespaceSeparator: '~',
-  reactNamespace: false,
-  defaultNamespace: 'plugin__console-plugin-template',
-  useKeysAsDefaultValue: true,
-
-  // see below for more details
-  lexers: {
-    hbs: ['HandlebarsLexer'],
-    handlebars: ['HandlebarsLexer'],
-
-    htm: ['HTMLLexer'],
-    html: ['HTMLLexer'],
-
-    mjs: ['JavascriptLexer'],
-    js: ['JavascriptLexer'], // if you're writing jsx inside .js files, change this to JsxLexer
-    ts: ['JavascriptLexer'],
-    jsx: ['JsxLexer'],
-    tsx: ['JsxLexer'],
-    json: [CustomJSONLexer],
-
-    default: ['JavascriptLexer'],
-  },
-};
