+++ package.json
@@ -15,7 +15,7 @@
     "build-dev": "yarn clean && rspack",
     "start": "rspack serve",
     "start-console": "./start-console.sh",
-    "i18n": "./i18n-scripts/build-i18n.sh && node ./i18n-scripts/set-english-defaults.js",
+    "i18n": "i18next-cli lint && i18next-cli extract",
     "test": "LANG=en_US.UTF-8 jest",
     "lint": "eslint src integration-tests --fix --max-warnings 0 && stylelint 'src/**/*.css' --allow-empty-input --fix",
     "test-e2e": "playwright test --headed",
@@ -53,11 +53,10 @@
     "eslint-plugin-testing-library": "^7.16.2",
     "globals": "^17.7.0",
     "i18next": "^25.8.18",
-    "i18next-parser": "^9.4.0",
+    "i18next-cli": "1.50.3",
     "jest": "^30.4.2",
     "jest-environment-jsdom": "^30.4.1",
     "jest-junit": "^17.0.0",
-    "pluralize": "^8.0.0",
     "prettier": "^3.9.6",
     "prettier-stylelint": "^0.4.2",
     "react": "^18.3.1",
